requirements.txt
aiohttp==3.7.3
aiokafka==0.7.0
asyncpg==0.21.0
protobuf==3.14.0
pytz==2020.4
pyyaml==5.3.1
yarl==1.6.3
